Feature/kr 75

+++ domain-order/src/main/java/co/dalicious/domain/order/converter/OrderTypeConverter.java
@@ -6,14 +6,14 @@ import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
 
 @Converter
-public class OrderTypeConverter implements AttributeConverter<OrderType, Long> {
+public class OrderTypeConverter implements AttributeConverter<OrderType, Integer> {
     @Override
-    public Long convertToDatabaseColumn(OrderType orderType) {
+    public Integer convertToDatabaseColumn(OrderType orderType) {
         return orderType.getCode();
     }
 
     @Override
-    public OrderType convertToEntityAttribute(Long dbData) {
+    public OrderType convertToEntityAttribute(Integer dbData) {
         return OrderType.ofCode(dbData);
     }
 }
